Enable wifi networking on first boot
The connman configuration in nixpkgs is somewhat wonky, but can be set
right.

connman itself also seems to suffer from a timing issue with RFKILL
blocked technologies. If first initialized while wifi hardware is
soft-blocked, it will disable the wifi technology and not reenable it
even when wpa_supplicant responds to a later unblock.

For reasons unknown, rfkill soft-blocks the built-in wifi for a split
second before immediately unblocking it. Restarting connman some seconds
after boot makes sure it reinitializes with wifi unblocked. Ugly.