Add localization GUI for timezone/lang/keymap
NixOS's systemd fork does not permit setting locale or keymap via
localed/localectl. Instead, we use the `LANG` env variable and
`setxkbmap` to configure locale for the xsession of the kiosk user. The
configuration is stored on a persistent folder.

The treatment of the timezone differs slightly, in that NixOS allows for
it to be controlled via timedated/timedatectl. Here we only add a
systemd service to set the timezone (system wide), based on a persistent
configuration file.