Defensively protect the kiosk connection daemon loop
Because we’re running in a daemon, some code may not be fully
initialised from the callback we get in the constructor.

There was this error: _set_captive_portal_url was not accessible.

This error has been fixed by starting the connection daemon after every
element had been fully initialized. But with future develop, it could
happen again.