Keep the same page when changing URL.
This fix a memory leak issue when changing page every time.

The reason we changed the page is that, if the page is slow to load,
from the moment we press the toggle shortcut:

- with a new page: a blank screen directly appears,
- with the same page: it keeps the current page until the new page is
received.

When using the same page, using `page.setContent(Widget())` or
`page.setHtml('')` in order to clear the screen before the new page
load does not have any effect.