Clean connection state after a connection failure
Previously, when an error happened during a connection, the agent was
not cleaned up. At the next connection attempt, an Already Existed error
was returned. A reboot of PlayOS was required to be able to connect to a
wifi network.

The proposed change clean up the agent, no matters if the connection
succeeds or fail. It allows to reconnect without restarting PlayOS.