Show service strength alongside its name in networks
The strength was already accessible in each service detail, but it was
too hidden to get this information at a glance.

When the strength of a network is too low, the user knows he will have
difficulties connecting to it. It can be a trigger for him to take
action to get a better strength if possible.