Increase BBS scans expiration
This change aims to fix WIFI connection when there are lot of available
WIFI networks.

When loading the network page, the list of WIFI networks are sent to the
user. When he wants to connect to one of them, he sends the id of the
WIFI network he wants to connect to, then, the server fetchs the list of
WIFI networks again and tries to find the user one. But when there are
lot of WIFI networks available, only a subset of available WIFI networks
are returned at each scan. It’s possible that the network the user wants
to connect to is not returned this time.

By default, WIFI networks are discarded if they don’t appear in 2
consecutive scans, see [wpa_supplicant configuration documentation](
https://w1.fi/cgit/hostap/plain/wpa_supplicant/wpa_supplicant.conf). We
could increase the number of required succesive scan before abandoning a
network.

After shutting down a WIFI network, I measured the time it was displayed
in cmst, according to various count of BBS expirations.

- 2 [default]: 25 seconds
- 4: 34 seconds
- 8: 52 seconds
- 16: 1 minute 30 seconds

By requiring 16 consecutive scans before droppinga network, that could
be enough to keep networks in a crowded WIFI place. The drawback is that
if a WIFI network is off, the user will know it 1mn30s after, instead of
25s after.