Add wpa_supplicant restart hook on rfkill unblock
This uses udev to restart wpa_supplicant whenever a rfkill block of wifi
is removed ("unblock"). This seems a bit trigger happy, but is probably
the most direct solution as long as the underlying problem in
wpa_supplicant/connman's interaction with rfkill is not known. Our
machines are also unlikely to toggle rfkill during operation, so the
practical downsides should be low.

While the heuristic timer-based restarting of wpa_supplicant mostly
works in regular setups, picking a timer duration is difficult:
Longer durations are more likely to work (coming after the unblock), but
overly long durations are bad for the user and may even lead to mistaken
conclusions of there being no connectivity at all.

Point in case, the duration was just about too short for live systems
("lab key"), so they would frequently boot into a state where wifi
connectivity could not be restored without a system reboot.