Add persistence of preferred boot order
The preferred boot order is persisted to preserve the boot choice after
manual selection from the GRUB menu. By relying on the boot order, the
manual choice will automatically be overridden following a successful
update of the inactive partition, and fallback booting continues to
work if the favoured choice fails repeatedly.