Patch rauc to use our notion of 'primary slot'
rauc assumes that a slot with a bad try is already "out" (on strike
rule). We have configured GRUB to use a three strike rule, however.

This patch makes rauc behave in line with our GRUB boot logic.

Patching seems to be the suggested way of configuring this:

> If you want to implement different behavior or use other variable
names, you might need to modify the uboot_set_state() and
uboot_set_primary() functions in src/bootchooser.c.

https://rauc.readthedocs.io/en/v1.2/integration.html