Fix installation
* On the replacement of /dev/root to /dev/disk/by-label/system.x on
/etc/fstab

I could not find the reason why this is failing now, but was ok before.
There may be a simpler solution than the current implemented one.

The problem already appears with NixOS 19.09. I looked at `systemd`
changelogs, it was upgraded from 239 to 243. In version 242:

    The .device units generated by systemd-fstab-generator and other
    generators do not automatically pull in the corresponding .mount
    unit as a Wants= dependency. This means that simply plugging in the
    device will not cause the mount unit to be started automatically.
    But please note that the mount unit may be started for other
    reasons, in particular if it is part of local-fs.target, and any
    unit which (transitively) depends on local-fs.target is started.

* On the VM graphics controller option

Using the modified value as before lead to a black screen, xserver can’t
be started. Using the default VirtualBox value shows 2 lines of errors,
but xserver starts.

* On /mnt/system rw instead of ro

The `nix` activation script needs to have access to `/nix/store`. This
change does not seem to be necessary for things to work, but it permits
to suppress errors.