Fix set-timezone service failures
The set-timezone service was repeated over and over, undefinitely.
Sometimes it succeeded, but most of the time, it failed.

Removing PathExists on set-timezone.path, and starting
set-timezone.service at startup is fixing the problem.

Because the service is run at startup, no matter if
/var/lib/gui-localization/timezone exists, we now have to check if this
file exists in set-timezone execution.

On develop, there is no problem. And PathExists is rigthly
triggered at startup if the file exist.

Looking at systemd documentation:

https://www.freedesktop.org/software/systemd/man/systemd.path.html

> When a service unit triggered by a path unit terminates (regardless whether
> it exited successfully or failed), monitored paths are checked immediately
> again, and the service accordingly restarted instantly. As protection against
> busy looping in this trigger/start cycle, a start rate limit is enforced on
> the service unit, see StartLimitIntervalSec= and StartLimitBurst= in
> systemd.unit(5). Unlike other service failures, the error condition that the
> start rate limit is hit is propagated from the service unit to the path unit
> and causes the path unit to fail as well, thus ending the loop.

It’s not clear whether `PathExists` is triggered unefinitely, and why it
did not cause any harm previously.