Fix localization state write
`Util.write_to_file` was writing the string to the file as a
replacement, not as a complete modification.

1. write Accra -> Accra (OK)
2. write Johannesburg -> Johannesburg (OK)
3. write Accra -> Accranesburg (KO)

Using the `O_TRUNC` flag truncate to 0 length if the file exists.
http://caml.inria.fr/pub/docs/manual-ocaml/libref/Unix.html#TYPEopen_flag

`Utils.write_to_file` is used on:

- /var/lib/gui-localization/timezone
- /var/lib/gui-localization/keymap
- /var/lib/gui-localization/lang

But only writes on the timezone file are affected, because keymaps and
langs all have the same length (see controller/server/gui.ml), hence the
bug does not appear there.