Reload the proxy on change in the kiosk
This will prevent users the need to restart their computers after a
modification of their proxy.

We get the initial proxy through dbus, and then, we listen to changes
through dbus as well.

To get some understanding, this command monitors dbus events in the
shell:

```bash
dbus-monitor --system "type='signal',sender='net.connman'"
```

There is a possibility that a proxy change is omitted initially, if it
happens after our initialization, but before the monitoring is active.