Use real time proxies in controller instead of fixed from startup
This goes along with using real time proxies in the kiosk.

Instead of using the proxy defined at startup, we get the proxy of the
currently connected service before doing any request in:

- update (status and download rauc bundle),
- network (check connection),
- gui overview (show proxy).

This will prevent using an undefined proxy at startup if it’s defined
with some delay after startup.