Remove connman-manual-proxy command
This command is no longer required, the kiosk and the controller have
their own method to get proxies.

One downside is that we have to manually make sure the same method is
used in both the controller and the kiosk to get the proxy. Currently,
we return the proxy of the service that is either ready or connected.