Get internet status on the fly on the network page
When updating the proxy, the internet state was not updated.

Internet state is currenlty updated:

- every x seconds (x depending if internet is connected or not),
- and if the list of services is modified in connman.

See network.ml.

Watching the services with connman does not listen to invidual property
changes according to doc/manager-api.tx:

> This signal will only be triggered when the sort order of the service
> list or the number of services changes. It will not be emitted if only a
> property of the service object changes. For that it is required to watch
> the PropertyChanged signal of the service object.

So, we would have to setup a watch on a service when it’s selected.

Or we could manually trigger an internet reload on proxy change.

But I think it’s even simpler to just get the current internet status on
the fly, and more so because it’s only used in the network page.