Generate HTML with tyxml
Experimentation went well, so here is the modification replacing all of
Mustache to tyxml. (https://github.com/dividat/playos/pull/63)

Support for `tyxml-ppx` is dropped, because the function-based approach
works well, and has less strange errors, like the one when it does not
work when we put newlines between some tags.

We now have a non limited language to generate templates, which will
makes it easier to work on HTML pages, particularly now that we are
investing more into PlayOS.