Use variable online check timeouts
The online check should time out quickly when first opening the network
page, but should allow for longer probing when explicitly checking for
connectivity.

The default timeout used for direct navigation to network page (200 ms)
is just above the threshold for "instant" feedback[1], but subjectively
just below the threshold for being annoyingly slow.

The timeout when clicking the refresh "button" is now longer and less
likely to produce false negatives.

[1]: https://www.nngroup.com/articles/response-times-3-important-limits/