Couple DNS and static IP configuration
There were a couple of disadvantages to allowing DNS configuration for
DHCP-configured connections:

1. It could be confusing: When configuring a static IP, it is also
necessary to configure DNS, but users may not be aware/forget.

2. It could facilitate abuse: When nameservers are obtained through
DHCP, control is with the network. When nameservers can only be
set up together with a static IP, more knowledge about the network
is needed to get a functioning configuration. When nameservers can be
altered in a DHCP-configured connection, it is relatively easy to swap
in a malignant nameserver while leaving everything else unchanged.

Since DNS configuration for DHCP-configured connections is not a
required feature currently, we optimize for ease of use and correctness.