Remove pcsclite pin on 1.8.23
This upgrade pcsclite to version 1.9.0, which is the version in our
nixpkgs 20.09 version.

Now, why it was required (thank’s Johannes):

- pcsclited runs as a daemon (server) that can talk to smart cards.

- pcsclite also provides a library that allows user-space programs to
talk to pcsclited; this library is usually dynamically linked.

- We statically link the pcsclite library in the driver, the idea was
that this makes the driver more standalone and portable.

- pcsclite changed the ABI between pcsclite library and server in an
incompatible way, which is not a problem when the library is dynamically
linked, since usually library and server change together on a system. It
is a problem when the pcsclite library is statically linked and the
daemon is a different, incompatible version.

This means the pcsclited daemon and library used by the driver need to
be compatible versions, or need to be dynamically linked.

When first including the driver in PlayOS, we referenced the released
(statically built) binary and pinned pcsclite to a compatible version.
If we now actually build the driver as part of PlayOS, we don't have to
do a static build and would then no longer need to pin pcsclite.

There is another card for figuring out how to proceed with the driver
releases generally:
https://trello.com/c/8tD8pDpu/899-future-release-model-for-driver