Add configurable patch to limit virtual terminals
In typical usage scenarios of PlayOS the virtual terminals on
Ctrl-Alt-FN are not useful (no user that can be logged into) and may
lead to confusion when accidentally switched to (especially terminal
12).

This patch blocks any but a few whitelisted virtual terminals from being
switched to directly from the Xserver session running the kiosk.

Xserver allows to disable VT switching entirely with the DontVTSwitch
option, but does not offer an option to disable specific VTs. We
therefore create a suitable patch for Xserver to enforce the whitelist.
Applying the patch in Xserver seems more robust than patching XKB
keymaps to disable the key combinations in question, which would need to
be done and tested for every locale.

The patch to the Xserver sources is minimal and targets a file that
hasn't changed much in recent years. At the time of patch creation, the
xorg package set is defined in
https://github.com/NixOS/nixpkgs/blob/21.11/pkgs/servers/x11/xorg/default.nix

Inspiration for patching nested attribute set xorg:
https://github.com/NixOS/nixpkgs/issues/11944