Fix webview dialog memory leak
When opening and closing System Settings continuously, we could see that
memory consumption of the kiosk application was growing fast, until the
kiosk became very slow because the system is using SWAP.

The reason is the parent of the webview in dialog was not specified when
creating the webview. So, the memory was not freed when the dialog was
closed.

Setting the dialog as the webview parent is fixing the leak. We can see
that after getting higher, the memory is sometimes going down (thanks to
some garbage collection ?).