Move internet status check to the frontend
The check was previously done in the backend. In order not to make the
client to wait for too long, we timed out the request after 0.2 seconds.
But in some cases, the request would have succeeded with just a few
extra time.

Instead, doing the request in the frontend, we don’t need a timeout. We
show a loader while the request is in progress. Then, we show
“Connected” if the result is success, “Not Connected” otherwise.