Show menu in a left sidebar
This permits to show all the entries in the same place, instead of
having some entries on the left, and others on the bottom right.

Each entry has its icon and label, which makes it clearer, especially
concerning the shutdown entry. Moreover, this special entry has been
moved on its own in the bottom of the screen.

I took the opportunity to introduce some constants in CSS.