Add check to avoid disclosing proxy passwords
As PlayOS systems are often semi-public, secrets entrusted to the system
should be protected from unwarranted access.

We should not allow users to keep the same proxy password when changing
the rest of the proxy configuration, as this could be used to reveal the
password to untrusted proxy servers on the same network or Internet.

This adds a check to fail update if keeping password was requested but
other proxy fields were changed.