Restore audio output from kiosk
We previously ran PulseAudio as system-wide daemon. The reasons are not
documented, but this may have been in order to more easily
configure/inspect audio from afar through another user's terminal
access. This has not been used in a long time.

Running in system-wide mode is not recommended[1] and not necessary for
the kiosk to be able to play sound. Since NixOS 22.05 audio output from
the kiosk no longer works when running in system-wide mode. We therefore
move to running the recommended per-user daemon.

I briefly considered migrating to PipeWire instead. However, there
currently seems to be an issue which leads to the kiosk losing the
ability to play audio after prolonged inactivity.

[1]: https://www.freedesktop.org/wiki/Software/PulseAudio/Documentation/User/WhatIsWrongWithSystemWide/