Set stateVersion to current release in each build
We started seeing warnings about `system.stateVersion` being unset in
various build targets.

For any of the one-off build targets (rescue, installer, live) this
value really does not matter, but it makes sense to set this to whatever
the system is built with.

For installed systems, the ideal value would be the version this system
was initially installed with. We do not know this at build time,
however. As only a very limited range of things is actually persisted
past a reboot, and in particular system state is mostly volatile,
NixOS's state version adaptations should not be ordinarily necessary. We
therefore also set the version the system is built with here. This is
also what has been filled in as a default in all previous releases.