Add service to recover rauc statusfile on boot
A simple oneshot service to recover interrupted statusfile updates.
Details are described in the code.

Note that this code has been written based on a relatively specific
record of a stuck machine that was successfully recovered. It has not
been possible to test the code directly on an affected system, as the
scenario is hard to trigger in a directed manner.

In particular, it is assumed but not experimentally confirmed that the
emptied status.ini file will be empty in the sense of
`test -s status.ini`.