Use combined sink to play audio on all outputs
We previously relied on the default behaviour of audio being routed to
HDMI outputs on the Shuttle DH110 series. This same behaviour is not
present on other hardware.

This patch should ensure that sound is output to the PCs default device
as well as any connected HDMI or DisplayPort outputs. It will only
include outputs already connected when the system is started.

A more robust and hotplugging friendly solution may be to rely on DRM
events from udev for detecting changes to monitor outputs, then querying
and activating connected audio devices with a small executable using
ALSA lib (instead of parsing `aplay` output).