Widen range of responses seen as captive
This changes captive portal detection to consider networks that replace
the probe URL response without a redirect.

We therefore check for a known text from the probe server to be
contained in the body if we receive a 200 (used by some Cisco networks
together with an HTML refresh), and switch to captive if it is missing.
In addition, some authentication related status codes are seen as
indicative of being captive.

We still refrain from interpreting just any unexpected response as being
captive, to avoid showing confusing prompts in case our probe server is
misconfigured, down or blocked.