Align overriding of nixpkgs.pkgs
Previously setting `nixpkgs.pkgs` twice (in base layer and live system
definition) worked, somehow. With nixpkgs for 23.05, this led to a
failure during evaluation:

    error: The option `nixpkgs.pkgs' is defined multiple times while it's expected to be unique.
       A Nixpkgs pkgs set can not be merged with another pkgs set.
       Definition values:
       - In `/nix/store/hl93xdyddynkvgpvfpvyvy1qvw0jvf11-source/nixos/default.nix'
       - In `/nix/store/hl93xdyddynkvgpvfpvyvy1qvw0jvf11-source/nixos/default.nix'
       Use `lib.mkForce value` or `lib.mkDefault value` to change the priority on any of these definitions.

We can avoid this issue by setting the override only in the base
definition, it does not need to be repeated in the live system
definition. Using `mkDefault` for the definition in the base layer as
recommended in the option's documentation[^1].

The installer system does not rely on the base layer and only refers to
it via an already built disk image. But we do not need to override the
`pkgs` property in it at all, so we leave the property unset.

[^1]: https://search.nixos.org/options?channel=23.05&show=nixpkgs.pkgs&from=0&size=50&sort=relevance&type=packages&query=nixpkgs.pkgs