Exclude installation medium as installation target
We try to exclude the installation medium as an auto-detected
installation target, to prevent issues in cases such as

- atypically large USB drive as installer medium,
- missing or undetected HDD due to misconfiguration.

To help understand the filtering process and possible failure scenarios
we now also list all the detected and the considered devices.

A note on the implementation: Detection of the boot device relies on
knowledge about the mount points, i.e. that the iso-image configuration
from Nixpkgs creates a mount point at `/iso`[^1]. This is the most
robust and straightforward mechanism I could think of, but is bound to
that mount point continuing to exist.

[^1]: https://github.com/NixOS/nixpkgs/blob/3c15feef7770eb5500a4b8792623e2d6f598c9c1/nixos/modules/installer/cd-dvd/iso-image.nix#L685-L689