Give access to captive portal when on settings
Previously, when the settings modal was open, when a captive portal was
detected, we showed a header with a button, but it could not be
interacted with, because the modal didn’t allow to access the parent
window.

Setup the settings modal to be modeless, so that the button to open the
captive portal is still accessible.

Also better represent the dialog state with an enumeration.

See https://trello.com/c/wcIrvEau/1445-captive-portal-prompt-pops-up-beneath-controller-modal