Set audio output to HDMI profile
The combined sink stopped working reliably after migrating to Nixpkgs
23.05 as a basis. While it could be "fixed" by directly playing a test
sound on the HDMI device first, I did not find a way to make the
combined sink work immediately after booting.

As the only standard way in which we output audio is the HDMI output, we
now instead explicitly set that output profile for the card.