Add workaround for Qt race condition
This works around the race condition described in
https://bugreports.qt.io/browse/QTBUG-111541 by waiting an instant after
triggering the clearing of profile cache. If a load is triggered while
the clearing is not finished, it can cause the load to hang. This
can be recovered from by triggering a simple reload, but is better to
avoid.

A signal for observing the progress of cache clearing should be
available for future versions of Qt 6:
https://github.com/qt/qtwebengine/commit/37430020d373304f812fa822651cf2a0ccf636d3

I considered adding a more involved mechanism to detect stalled loads
and automatically trigger a new load, but concluded that a sleep should
be overall simpler and easy enough to understand. As a relatively short
sleep is chosen to avoid user confusion, this may still fail, but a
hanging load is relatively harmless and should be recoverable by the
user.