Map most system control keys to poweroff
With the typical use of the kiosk system there is little point to
suspending the machine, and it is more likely to confuse users than
help. With this change we set most simple key presses to ignored and map
all long presses to poweroff, which is the sensible action.

Because at this level both keyboard power keys and "regular" on-device
power keys are handled, we can not disable the power key completely.

When expected remote control devices have been clarified, we may want to
go back and handle all simple presses in the UI instead of ignoring or
immediately action on them, asking for a second press to confirm
poweroff. In this case logind can be told to hold off via DBUS locks.

As an alternative I tried resetting the key mappings via `xmodmap`, but
found that the keys still seemd to be handled at the low level.

https://www.freedesktop.org/software/systemd/man/latest/logind.conf.html#HandlePowerKey=