Inhibit short press on remote control power off
Manage short press upon power off ourselves, and not with systemd
anymore, so that we only listen to the Power Button, and not to power
off buttons located on keyboards or remote controls.

This means pressing the Power Button on the computer with a short press
directly shutdowns the computer, and pressing power off on a remote
control with a short press does nothing.