Add firewall rule for TFTP conntracking
This adds a firewall output rule to allow use of a TFTP client for Senso
firmware updates. In the TFTP protocol, for a write request, the client
selects a random local port (pclient) and makes a request to the
server's port 69, which chooses a random port which it responds from
(pserver) and sends an ACK to the client's chosen port. The rest of the
packets are sent exclusively between those two randomly chosen ports,
and port 69 is out of the picture. In order for the client-side firewall
to accept the server's responses at the random port, stateful connection
tracking is required to temporarily expect inbound packets from the
known peer. TFTP builds on session-less UDP, so a protocol-specific
helper is needed to inspect the TFTP packets and set up the correct
expectation (inbound from pserver to pclient, instead of the naive
expectation of inbound from 69 to pclient).

We register the tftp conntrack helper explicitly and only for outbound
connections (i.e. allowing only TFTP _clients_).

Due to terrible and often outdated documentation, the crucial line was
surprisingly difficult to come up with. The following commands/tools and
references were helpful in debugging the configuration:

- `lsmod | grep conntrack`
- `conntrack -E` with `--dst` filter (from contrack-tools)
- `iptables -t raw -L` / `nft list ruleset`
- WireShark
- https://wiki.nftables.org/wiki-nftables/index.php/Conntrack_helpers

The patch is written to stick to the iptables frontend to netfilter,
which is currently the default in NixOS. The nftables equivalent (also
tested and confirmed to work) would be this:

    networking.firewall = {
      enable = true;

      connectionTrackingModules = [ "tftp" ];
    };
    networking.nftables.enable = true;
    networking.nftables.tables.nixos-fw.content = ''
      ct helper tftp-69 {
        type "tftp" protocol udp;
      }

      chain output {
        type filter hook output priority filter;
        udp dport 69 ct helper set "tftp-69"
      }
    '';

Support for nftables is maturing in NixOS, but the automatic
equivalent configuration to the iptables (frontend) default is still
relatively fresh, so sticking to the iptables frontend for now in order
to avoid wider retesting of connectivity across different services.
Maybe with the next Nixpkgs update the transition has smoothened out a
bit more upstream and we can switch to nftables completely.