Use updated driver with fw update support
Use a snapshot version of the driver with support for firmware updates
triggered from the application.

With this version of the driver, connections from origins other than the
whitelisted permissible origins are actively rejected instead of just
omitting the CORS headers. We set the base URL of the kiosk URL as the
only permissible origin.

Vendor hash updated by swapping in `lib.fakeHash` to have updated hash
be revealed by Nix.