Allow using the kiosk on platforms without DBus
We previously added support for using the kiosk without connman
installed, but it still crashed when started on systems without DBus
(i.e. any non-Linux platform).

With this change we add a stub implementation of the core proxy
abstraction, which is going to provide safe no-op responses on non-Linux
platforms, allowing to use the kiosk, albeit without proxy support.