Fix `nix-shell --run`
Executing another bash inside the shell hook seems to break running
commands with `nix-shell --run`, something we assumed to work in the CI
configuration.

It seems that the Qt wrapper is already set up at least in modern
Nixpkgs, so we can simply remove the code for spawning the wrapped bash.