Fix kiosk hard refresh
Trying to hard refresh led to the kiosk being stuck with the loader. To
fix this, this was required to apply a simple refresh instead.

In order to fix this, use the refresh timer, instead of using
time.sleep.

Once we upgrade Qt to the 6.7 version, in next release of nixpkgs, we’ll
be able to use the provided signal, that activates once the cache has
been cleared.