Fix focus after reload
Since the migration to Qt6, reloading the browser, either using CTRL+R,
or CTRL+SHIFT+R, the focus is lost on the webview.

By clearing the focus, it’s properly set afterward, after a reload.

I also tried to play with focus policy, and setting focus after a bit
with a timer, but without success.