Add basic integration test for kiosk proxy with connman
This is the humble beginnings of an automated integration test for proxy
use in the kiosk.

At the moment the test is rather limited: It only checks that a proxy
configured with connman is picked up and set as application proxy by the
kiosk.

Unfortunately I have not been able to get the proxy to be actually used
by the Qt framework in this test setup. It rather seems like the proxy
is properly configured from our side. It would be nice to extend the
tests to confirm that requests with the proxy can successfully be made,
but this may require analysis of the Qt framework in the context of the
VM test setup.

I am including two sanity checks that the basic node topology is as
expected (no direct route from client to 'the cloud'), and that the
proxy node basically works (with curl).