Update build/quick start instructions
The quick start instructions were actually more like "slow start"
instructions, suggesting to do a full build. Even worse, `nix build`
fails in modern versions of the Nix CLI, which expects to find a flake
file with this invocation.

This suggests to use the commonly used `./build` script instead and
updates it to provide a minimal bit of guidance when invoking it empty
(instead of doing, again, the full build).

A full build can still be done by invoking `./build all`.