Use tagged version of focus-shift
This indicates more in the face which version of focus-shift we are
using.

This also upgrade `focus-shift` with the following change:

- Ignore non-rendered elements

We could think about using `nix` to pull the dependency, but as this may
not be updated that much, and as we have only this dependency, keep it
simple for now?