Use an on-the-record profile named 'Default'
In Qt 6 the default behaviour was changed to create off-the-record
profiles, which are not written to disk. In order to obtain persistent
profiles again, we explicitly create a profile named "Default", which
matches the default naming in Qt 5.

We unset the page when the app is closed, to avoid the following warning
due to page and profile being torn down in the wrong order:

    Release of profile requested but WebEnginePage still not deleted. Expect troubles !