Set an explicit app name for kiosk-browser
We previously let the app name be inferred from the name of the
executable, but this is brittle and caused a rename when the Nix Qt
wrapper configuration was changed.

We explicitly set the app name to "kiosk-browser", which is what was
the previously inferred name with the wrapped executable.

This is important to preserve earlier cache and storage locations which
are derived from the app name, see
https://doc.qt.io/qt-6/qstandardpaths.html