Perform factory reset/wipe via stage-1 postDeviceCommands
We exploit the fact that /boot is already mounted as writable in normal
PlayOS installations and used to persist critical system health/status
data (i.e. RAUC's /boot/status.ini, grubenv).

Keeping the systemd service wrapper to avoid changing the rest of the
implementation and it is perhaps a good idea to have it as a general
interface to this functionality.